fix: refactor api file ad bump libraries
