+++ app/layout.tsx
@@ -1,8 +1,10 @@
 import "../styles/globals.css";
 import { getServerSession } from "next-auth";
+
 import SessionProvider from "@components/SessionProvider";
 import Login from "@components/Login";
-import { authOptions } from "@api/auth/[...nextauth]";
+import { authOptions } from "@api/auth/[...nextauth]/route";
+
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
