+++ package.json
@@ -12,29 +12,27 @@
     "prettier:write": "prettier --write ."
   },
   "dependencies": {
-    "@heroicons/react": "^2.0.16",
+    "@heroicons/react": "^2.0.18",
     "@szum-tech/design-system": "^1.5.5",
-    "firebase": "^9.17.2",
-    "firebase-admin": "^11.5.0",
-    "next": "^13.2.3",
-    "next-auth": "^4.20.1",
-    "openai": "^3.2.1",
+    "firebase": "^10.1.0",
+    "firebase-admin": "^11.10.1",
+    "next": "^13.4.16",
+    "next-auth": "^4.23.0",
+    "openai": "^3.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-firebase-hooks": "^5.1.1"
   },
   "devDependencies": {
-    "@tailwindcss/container-queries": "^0.1.0",
-    "@types/node": "^18.14.6",
-    "@types/prettier": "^2.7.2",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "autoprefixer": "^10.4.12",
-    "postcss": "^8.4.18",
-    "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.4",
-    "tailwind-scrollbar": "^2.1.0",
-    "tailwindcss": "^3.2.7",
-    "typescript": "^4.9.5"
+    "@types/node": "^20.5.0",
+    "@types/prettier": "^3.0.0",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
+    "autoprefixer": "^10.4.15",
+    "postcss": "^8.4.27",
+    "prettier": "^3.0.2",
+    "prettier-plugin-tailwindcss": "^0.5.3",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.1.6"
   }
 }
